WIP Library factor out repeated components
